+++ pygapit/io/_storage_numpy.py
@@ -5,14 +5,8 @@ from __future__ import annotations
 import json
 import typing as t
 from pathlib import Path
-from sys import version_info
 from types import TracebackType
-from typing import TypedDict
-
-if version_info >= (3, 11):
-    from typing import Self
-else:
-    from typing_extensions import Self
+from typing import Self, TypedDict
 
 import numpy as np
 import pandas as pd